cleanup graphql query names and fragments
