+++ client/src/features/packs/PackCreatorPage.tsx
@@ -1,179 +1,13 @@
 import { useState } from "react";
 import styled from "styled-components";
-import { Link } from "react-router-dom";
 import { useParams } from "react-router-dom";
-import { gql, useQuery, useMutation } from "@apollo/client";
-import { useAlert } from "react-alert";
-import ReactTooltip from "react-tooltip";
+import { gql, useQuery } from "@apollo/client";
 
-import { useHostGame } from "features/game/gameService";
-
+import { Topbar } from "features/packs/components/Topbar";
 import { Sidebar } from "features/packs/components/Sidebar";
 import { ScenePreview } from "features/packs/components/ScenePreview";
-import { NavigationContainer } from "features/packs/components/Navigation";
-import { PackForm } from "features/packs/components/PackForm";
-import { Button, Icon, Modal, Loader } from "components";
 
-import {
-  PackCreatorPagePackQuery,
-  PackCreatorPagePackQuery_pack,
-} from "./__generated__/PackCreatorPagePackQuery";
-import { PackUpdateMutation } from "./__generated__/PackUpdateMutation";
-
-export const PACK_FRAGMENT = gql`
-  fragment PackSettingsFragment on Pack {
-    id
-    name
-    description
-    length
-    isRandom
-  }
-`;
-
-const PACK_UPDATE = gql`
-  mutation PackUpdateMutation($input: PackUpdateInput!) {
-    packUpdate(input: $input) {
-      pack {
-        ...PackSettingsFragment
-      }
-    }
-  }
-  ${PACK_FRAGMENT}
-`;
-
-export const Navigation = ({
-  pack,
-  saving,
-  setSaving,
-}: {
-  pack: PackCreatorPagePackQuery_pack;
-  saving: boolean;
-  setSaving: (saving: boolean) => void;
-}) => {
-  const alert = useAlert();
-  const [isOpen, setIsOpen] = useState(false);
-  const [packUpdate] = useMutation<PackUpdateMutation>(PACK_UPDATE);
-  const hostGame = useHostGame();
-
-  const onSaveChanges = async (newPackInfo = {}) => {
-    setSaving(true);
-    const newPack = { ...pack, ...newPackInfo };
-    try {
-      await packUpdate({
-        variables: {
-          input: {
-            id: newPack.id,
-            name: newPack.name,
-            is_random: newPack.isRandom,
-            length: newPack.length,
-          },
-        },
-      });
-      setSaving(false);
-      setIsOpen(false);
-    } catch (error) {
-      alert.show(error.message);
-      setSaving(false);
-    }
-  };
-
-  return (
-    <StyledNavigationContainer>
-      <div className="left">
-        <Link to={`/packs/${pack.id}`}>
-          <img className="logo" src="/logo/logomark.svg" alt="Playhouse" />
-        </Link>
-      </div>
-      <div className="right">
-        <div>
-          <Loader loading={saving} />
-        </div>
-        <button onClick={() => setIsOpen(true)}>
-          <h4 className="pack-title">{pack?.name}</h4>
-        </button>
-        <div>
-          <ReactTooltip effect="solid" place="bottom" />
-          <Button
-            className="pack-ext-button"
-            variant="fab"
-            onClick={() => setIsOpen(true)}
-            data-tip="Edit pack"
-          >
-            <Icon icon="pencil" />
-          </Button>
-          <Button
-            className="pack-ext-button"
-            variant="fab"
-            onClick={() => hostGame(pack.id)}
-            data-tip="Test play"
-          >
-            <Icon icon="play" />
-          </Button>
-        </div>
-      </div>
-      <Modal
-        open={isOpen}
-        title="Pack Settings"
-        onRequestClose={() => setIsOpen(false)}
-        maxWidth={500}
-        closeButton
-      >
-        <PackForm
-          onSubmit={onSaveChanges}
-          loading={saving}
-          submitText="Save Changes"
-          defaultValues={pack}
-        />
-      </Modal>
-    </StyledNavigationContainer>
-  );
-};
-
-const StyledNavigationContainer = styled(NavigationContainer)`
-  .left {
-    ${({ theme }) => theme.media.desktop`
-      border-right-color: ${theme.ui.backgroundInverse};
-    `}
-  }
-  .right {
-    position: relative;
-  }
-  .loader {
-    position: absolute;
-    left: ${({ theme }) => theme.spacings(4)};
-    top: ${({ theme }) => theme.spacings(4)};
-    color: ${({ theme }) => theme.ui.lightText};
-  }
-  .pack-ext-button {
-    margin-right: ${({ theme }) => theme.spacings(2)};
-  }
-  .pack-ext-button:last-child {
-    margin-right: 0;
-  }
-`;
-
-const PACK_QUERY = gql`
-  query PackCreatorPagePackQuery($packId: ID!, $sceneId: ID) {
-    pack(id: $packId) {
-      ...PackSettingsFragment
-      ...SidebarPackFragment
-    }
-    scene(id: $sceneId, packId: $packId) {
-      ...ScenePreviewFragment
-    }
-    questionTypes {
-      id
-      slug
-    }
-    answerTypes {
-      id
-      slug
-    }
-  }
-  ${PACK_FRAGMENT}
-  ${Sidebar.fragments.pack}
-  ${ScenePreview.fragments.scene}
-`;
+import { PackCreatorPagePackQuery } from "./__generated__/PackCreatorPagePackQuery";
 
 export const PackCreatorPage = () => {
   const [saving, setSaving] = useState(false);
@@ -196,7 +30,7 @@ export const PackCreatorPage = () => {
     <Page>
       {data?.pack && (
         <>
-          <Navigation pack={data.pack} saving={saving} setSaving={setSaving} />
+          <Topbar pack={data.pack} saving={saving} setSaving={setSaving} />
           <Sidebar
             pack={data.pack}
             selectedSceneId={data.scene?.id}
@@ -214,6 +48,29 @@ export const PackCreatorPage = () => {
     </Page>
   );
 };
+
+const PACK_QUERY = gql`
+  query PackCreatorPagePackQuery($packId: ID!, $sceneId: ID) {
+    pack(id: $packId) {
+      ...TopbarPackFragment
+      ...SidebarPackFragment
+    }
+    scene(id: $sceneId, packId: $packId) {
+      ...ScenePreviewFragment
+    }
+    questionTypes {
+      id
+      slug
+    }
+    answerTypes {
+      id
+      slug
+    }
+  }
+  ${Topbar.fragments.pack}
+  ${Sidebar.fragments.pack}
+  ${ScenePreview.fragments.scene}
+`;
 
 export const Page = styled.section`
   height: var(--vh, 100vh);
